Move GCPU to the gcpu directory since names must match
